refactor: isomorphic snapshot for CLI (#3728)
